+++ packages/cinterop/src/jvm/jni/java_class_global_def.hpp
@@ -63,6 +63,7 @@ private:
         , m_io_realm_kotlin_internal_interop_sync_async_open_callback(env, "io/realm/kotlin/internal/interop/AsyncOpenCallback", false)
         , m_io_realm_kotlin_internal_interop_progress_callback(env, "io/realm/kotlin/internal/interop/ProgressCallback", false)
         , m_io_realm_kotlin_internal_interop_app_callback(env, "io/realm/kotlin/internal/interop/AppCallback", false)
+        , m_io_realm_kotlin_internal_interop_connection_state_change_callback(env, "io/realm/kotlin/internal/interop/ConnectionStateChangeCallback", false)
     {
     }
 
@@ -86,6 +87,7 @@ private:
     jni_util::JavaClass m_io_realm_kotlin_internal_interop_sync_async_open_callback;
     jni_util::JavaClass m_io_realm_kotlin_internal_interop_progress_callback;
     jni_util::JavaClass m_io_realm_kotlin_internal_interop_app_callback;
+    jni_util::JavaClass m_io_realm_kotlin_internal_interop_connection_state_change_callback;
 
     inline static std::unique_ptr<JavaClassGlobalDef>& instance()
     {
@@ -196,6 +198,11 @@ public:
     inline static const jni_util::JavaClass& app_callback()
     {
         return instance()->m_io_realm_kotlin_internal_interop_app_callback;
+    }
+
+    inline static const jni_util::JavaClass& connection_state_change_callback()
+    {
+        return instance()->m_io_realm_kotlin_internal_interop_connection_state_change_callback;
     }
 
     inline static const jni_util::JavaMethod function0Method(JNIEnv* env) {